Restructured silk creating separate mod
