+++ symphonia-codec-opus/src/celt/decoder.rs
