+++ symphonia-codec-opus/src/celt/mod.rs
@@ -1,0 +1,3 @@
+mod decoder;
+
+pub (crate) use decoder::Decoder;
