+++ symphonia-codec-opus/src/entropy.rs
