+++ symphonia-codec-opus/src/silk/filter/mod.rs
@@ -1,0 +1,5 @@
+mod lsf;
+mod lpc;
+mod ltp;
+mod pitch;
+mod prefilter;
