+++ symphonia-codec-opus/src/silk/mod.rs
@@ -1,0 +1,10 @@
+mod filter;
+mod frame;
+mod vq;
+mod excitation;
+mod table;
+mod codebook;
+mod error;
+mod decoder;
+
+pub (crate) use decoder::Decoder;
